Update sync
When importing the project into Android Studio, these changes were
necessary to make the sync happen without errors or warnings. It seems
like most of it is tool version updates.

Specifically for the changes from 'compile' to 'api' in
app/build.gradle: I'm not sure 'api' is necessary over the now
available 'implementation' for each dependency. The docs said that
'api' was equivalent to 'compile' so I went with 'api' out of an
abundance of caution. It's possible, perhaps even likely, that some or
all of these deps could be listed as 'implementation', which the docs
indicate could significantly decrease build time. But that's an
optimization for another day.