Separate app for debugging
Add .debug to appid and prefix D- to app name when running debug
builds. This allow the debug version of the app to live alongside the
production version on a single device. This is important because,
while I'm developing on this app, I'm also a user, and I'd rather
leave my useful install alone while I'm developing so that I can still
use it.